requirements.txt

# Django and core dependencies
Django==4.2.10
djangorestframework==3.14.0

# Google API Integration
google-api-python-client==2.126.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
requests==2.31.0

# Face Detection and Recognition
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# Database
psycopg2-binary==2.9.9

# UI
django-crispy-forms==2.1

# Development and deployment
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
